requirements.txt
tiktoken
langchainhub
langchain==0.2.8 
langgraph==0.0.64

langchain-community==0.2.3
python-dotenv
tavily-python 
langchain-text-splitters
langchain-openai==0.1.8
pandas==2.2.2
praw==7.7.1

fuzzywuzzy==0.18.0

streamlit==1.31.1
ollama==0.2.1
beautifulsoup4==4.12.3
termcolor==2.4.0

colorlog==6.8.2
fake-useragent==1.5.1
playwright==1.45.0

fsspec
googlesearch-python